rename discovery to public organizations, index and show
