--- app/controllers/discovers_controller.rb
@@ -1,5 +1,0 @@
-class DiscoversController < ApplicationController
-  def show
-    @organizations = Organization.not_privacy_setting_private
-  end
-end
