+++ app/views/public/organizations/_organization.html.erb
@@ -1,0 +1,25 @@
+<%= turbo_frame_tag "organization_#{organization.id}" do %>
+  <%= render CardComponent.new(title: organization.name, description: "Members: #{organization.users.count}") do |card| %>
+    <% card.with_image do %>
+      <%= link_to public_organization_path(organization), data: { turbo_frame: "_top" } do %>
+        <%= organization_avatar(organization, classes: "max-w-full h-80 object-contain") %>
+      <% end %>
+    <% end %>
+    <% card.with_action do %>
+      <% if organization.participant?(current_user) %>
+        <%= link_to t("public.organizations.show.actions.open"), organization_dashboard_path(organization), class: "du-btn du-btn-primary", data: { turbo_frame: "_top" } %>
+      <% elsif organization.user_requests.pending.exists?(user: current_user) %>
+        <%= link_to t("public.organizations.show.actions.waiting_for_approval"), "#", class: "du-btn du-btn-primary du-btn-disabled" %>
+      <% elsif organization.user_invitations.pending.exists?(user: current_user) %>
+        <%= button_to t("public.organizations.show.actions.accept_invitation"), approve_user_invitation_path(organization.user_invitations.pending.find_by(user: current_user)),
+          method: :post,
+          class: "du-btn du-btn-primary",
+          data: { turbo: false } %>
+      <% else %>
+        <%= button_to t("public.organizations.show.actions.#{organization.privacy_setting_public? ? "join" : "request_to_join"}"), user_membership_requests_path(organization_id: organization.id),
+          method: :post,
+          class: "du-btn du-btn-primary" %>
+      <% end %>
+    <% end %>
+  <% end %>
+<% end %>
