+++ app/views/public/organizations/index.html.erb
@@ -7,7 +7,7 @@
 
   <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
     <% @organizations.each do |organization| %>
-      <%= render "discovers/organization", organization: organization %>
+      <%= render "public/organizations/organization", organization: organization %>
     <% end %>
   </div>
 <% end %>
