+++ app/views/public/organizations/show.html.erb
@@ -1,0 +1,5 @@
+<% content_for :title, t(".title") %>
+
+<%= render PageComponent.new(title: @organization.name) do |component| %>
+  <%= render "public/organizations/organization", organization: @organization %>
+<% end %>
