+++ app/views/shared/_footer_nav.html.erb
@@ -10,7 +10,7 @@
       <span class="du-dock-label">Projects</span>
     <% end %>
   <% else %>
-    <%= active_link_to discover_path, active: :exclusive, class_active: "du-dock-active", class: "" do %>
+    <%= active_link_to public_organizations_path, active: :exclusive, class_active: "du-dock-active", class: "" do %>
       🧭
       <span class="du-dock-label">Discover</span>
     <% end %>
